+++ src/Accounts/Authenticators/ClientAssertionAuthenticator.cs
@@ -43,8 +43,12 @@ namespace Microsoft.Azure.PowerShell.Authenticators
             var options = new ClientAssertionCredentialOptions()
             {
                 AuthorityHost = new Uri(authority),
-                TokenCachePersistenceOptions = spParameters.TokenCacheProvider.GetTokenCachePersistenceOptions()
             };
+            if (!(spParameters.TokenCacheProvider.GetTokenCachePersistenceOptions() is InMemoryTokenCacheOptions))
+            {
+                options.TokenCachePersistenceOptions = spParameters.TokenCacheProvider.GetTokenCachePersistenceOptions();
+            }
+
             options.Diagnostics.IsTelemetryEnabled = false; // disable telemetry to avoid error thrown from Azure.Core that AssemblyInformationalVersion is null
             TokenCredential tokenCredential = new ClientAssertionCredential(tenantId, spParameters.ClientId, () => GetClientAssertion(spParameters), options);
             string parametersLog = $"- ClientId:'{spParameters.ClientId}', TenantId:'{tenantId}', ClientAssertion:'***' Scopes:'{string.Join(",", scopes)}'";